refactor(presentation): introduce domain route builders for garden
Closes #33
Split GardenAdvisor route mapping into domain-oriented route builders and
keep the existing public route behavior stable by leaving non-planning
slices as no-op boundaries for upcoming endpoint re-homing issues.
